client/src/requirements.txt
flask
grpcio
grpcio-tools
jsonpickle
protobuf
prometheus-client
prometheus_flask_exporter
requests